fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@24272 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/guisim/GUIPerson.cpp
@@ -227,7 +227,7 @@ GUIParameterTableWindow*
 GUIPerson::getTypeParameterWindow(GUIMainWindow& app,
                                   GUISUMOAbstractView&) {
     GUIParameterTableWindow* ret =
-        new GUIParameterTableWindow(app, *this, 8 + myVType->getParameter().getMap().size());
+        new GUIParameterTableWindow(app, *this, 8 + (int)myVType->getParameter().getMap().size());
     // add items
     ret->mkItem("Type Information:", false, "");
     ret->mkItem("type [id]", false, myVType->getID());
