+++ sumo/src/guisim/GUIVehicle.cpp
@@ -186,7 +186,7 @@ GUIParameterTableWindow*
 GUIVehicle::getTypeParameterWindow(GUIMainWindow& app,
                                    GUISUMOAbstractView&) {
     GUIParameterTableWindow* ret =
-        new GUIParameterTableWindow(app, *this, 22 + myType->getParameter().getMap().size());
+        new GUIParameterTableWindow(app, *this, 22 + (int)myType->getParameter().getMap().size());
     // add items
     ret->mkItem("Type Information:", false, "");
     ret->mkItem("type [id]", false, myType->getID());
