+++ sumo/src/mesogui/GUIMEVehicle.cpp
@@ -139,7 +139,7 @@ GUIParameterTableWindow*
 GUIMEVehicle::getTypeParameterWindow(GUIMainWindow& app,
                                      GUISUMOAbstractView&) {
     GUIParameterTableWindow* ret =
-        new GUIParameterTableWindow(app, *this, 9 + myType->getParameter().getMap().size());
+        new GUIParameterTableWindow(app, *this, 9 + (int)myType->getParameter().getMap().size());
     // add items
     ret->mkItem("Type Information:", false, "");
     ret->mkItem("type [id]", false, myType->getID());
